fix: address Gavin's audit findings (issues 1-5, 7)
- Fix 1: IShadow.PublicInput.chainId uint256->uint64; update
  ChainIdMismatch error signature to (uint64, uint64) to match the
  8-byte LE journal encoding used by the ZK circuit

- Fix 2: Add chainId == block.chainid guard in ShadowVerifier.verifyProof
  for defense-in-depth; previously only Shadow.claim enforced this

- Fix 3: Reject non-canonical hex-prefix even-path encodings in
  decode_compact_nibbles (padding nibble must be zero)

- Fix 4: Reject leading-zero (non-minimal) RLP integers in
  parse_u64_from_rlp_quantity

- Fix 5: Add immutable maxClaimAmount to Shadow (initialized to 8 ether,
  matching circuit MAX_TOTAL_WEI); claim() enforces the cap on-chain as
  a guardrail against verifier misconfiguration

- Fix 6 (doc): Document that the Shadow proxy owner must be the Taiko DAO
  governance contract; expand initialize() NatSpec accordingly

- Fix 7: Strict length validation in legacy_to_input; change < to != for
  amounts/recipientHashes/proofNodes to reject oversized arrays

+++ packages/contracts/test/Risc0CircuitVerifier.t.sol
@@ -220,7 +220,7 @@ contract Risc0CircuitVerifierTest is Test {
     function _sampleInput() private view returns (IShadow.PublicInput memory) {
         return IShadow.PublicInput({
             blockNumber: 4_353_615,
-            chainId: block.chainid,
+            chainId: uint64(block.chainid),
             amount: 1_230_000_000_000,
             recipient: 0xA92C80B3962F10e063Ad5463f996fe414F0E1F66,
             nullifier: keccak256("nullifier")
